// .agent/skills/umbraco-backoffice/examples/notes-wiki/Client/src/api/types.gen.ts
export type TreeItemModel = {
    id: string;
    name: string;
    entityType: string;
    hasChildren: boolean;
    isFolder: boolean;
    icon: string;
    parentId?: string | null;
};